backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
billiard==4.2.2
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.11.12
    # via
    #   cloudinary
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cloudinary==1.44.1
    # via
    #   -r requirements.in
    #   django-cloudinary-storage
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via -r requirements.in
dj-database-url==3.0.1
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2.8
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-allauth==65.13.0
    # via -r requirements.in
django-cloudinary-storage==0.3.0
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-yasg==1.21.11
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
joblib==1.5.2
    # via scikit-learn
kombu==5.5.4
    # via celery
numpy==2.3.4
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via
    #   drf-yasg
    #   gunicorn
    #   kombu
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via
    #   -r requirements.in
    #   reportlab
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   drf-yasg
    #   pandas
pyyaml==6.0.3
    # via drf-yasg
redis==7.0.1
    # via -r requirements.in
reportlab==4.4.4
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   django-cloudinary-storage
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.3
    # via scikit-learn
six==1.17.0
    # via
    #   cloudinary
    #   python-dateutil
sqlparse==0.5.3
    # via django
threadpoolctl==3.6.0
    # via scikit-learn
tzdata==2025.2
    # via
    #   django
    #   kombu
    #   pandas
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.5.0
    # via
    #   cloudinary
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
whitenoise==6.11.0
    # via -r requirements.in
